+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 - Add samples_count query to check how large is an index.
 - Fix http dir query not using the provided certificate.
+- Try to use the default /etc/pki/tls/certs/ca-bundle.crt CA for TLS verification.
 
 web-0.2.4 (next)
 =========